lxml: better error message
Since version 2.0, lxml comes with a dedicated Python package for dealing with
HTML: lxml.html. It is based on lxml's HTML parser, but provides a special
Element API for HTML elements, as well as a number of utilities for common HTML
processing tasks.

Since we're using lxml.html, if it isn't available, warn users to install it
instead of telling them to instal lxml.

Closes: #2